+++ test/tracking.jl
@@ -52,7 +52,7 @@ end
 
     tf = FrankWolfe.TrackingObjective(f,0)
     tgrad! = FrankWolfe.TrackingGradient(grad!,0)
-    tlmo = FrankWolfe.TrackingLMO(lmo_prob)
+    tlmo = FrankWolfe.TrackingLMO(lmo)
 
     x0 = FrankWolfe.compute_extreme_point(tlmo, spzeros(1000))
     storage = []
@@ -67,7 +67,7 @@ end
         max_iteration=5000,
         trajectory=true,
         callback=callback,
-        verbose=false,
+        verbose=true,
     )
 
     @test length(storage[1]) == 8
@@ -96,4 +96,34 @@ end
     @test tgrad!.counter == niters + 1
     # lazification
     @test tlmo.counter < niters
+end
+
+@testset "Testing vanilla Frank-Wolfe with various step size and momentum strategies" begin
+    f(x) = norm(x)^2
+
+    function grad!(storage, x)
+        @. storage = 2x
+    end
+    
+    lmo = FrankWolfe.ProbabilitySimplexOracle(1)
+
+    tf = FrankWolfe.TrackingObjective(f,0)
+    tgrad! = FrankWolfe.TrackingGradient(grad!,0)
+    tlmo = FrankWolfe.TrackingLMO(lmo)
+
+    x0 = FrankWolfe.compute_extreme_point(tlmo, spzeros(1000))
+    storage = []
+    callback = FrankWolfe.tracking_trajectory_callback(storage)
+
+    FrankWolfe.frank_wolfe(
+        tf,
+        tgrad!,
+        tlmo,
+        x0,
+        line_search=FrankWolfe.Agnostic(),
+        max_iteration=5000,
+        trajectory=true,
+        callback=callback,
+        verbose=false,
+    )
 end
